+++ pycognaize/document/snapshot.py
@@ -42,7 +42,8 @@ class Snapshot:
     def download(cls, snapshot_id: str, destination_dir: str,
                  exclude_images: bool = False,
                  exclude_ocr: bool = False,
-                 exclude_pdf: bool = False
+                 exclude_pdf: bool = False,
+                 exclude_html: bool = False
                  ) -> Tuple['Snapshot', str]:
         """Downloads snapshot to specified destination"""
         login_instance = Login()
@@ -52,9 +53,10 @@ class Snapshot:
                                          snapshot_id)
 
             exclude = cls._get_exclude_patterns(
-                exclude_images,
-                exclude_ocr,
-                exclude_pdf
+                exclude_images=exclude_images,
+                exclude_ocr=exclude_ocr,
+                exclude_pdf=exclude_pdf,
+                exclude_html=exclude_html
             )
 
             downloader = SnapshotDownloader()
@@ -91,7 +93,13 @@ class Snapshot:
         return snapshot_path
 
     @classmethod
-    def _get_exclude_patterns(cls, exclude_images, exclude_ocr, exclude_pdf):
+    def _get_exclude_patterns(
+            cls,
+            exclude_images,
+            exclude_ocr,
+            exclude_pdf,
+            exclude_html
+    ):
         exclude = []
 
         if exclude_images:
@@ -102,6 +110,9 @@ class Snapshot:
 
         if exclude_pdf:
             exclude.append('*.pdf')
+
+        if exclude_html:
+            exclude.append('*.html')
 
         return exclude